Создал папки Data, Domain, Presentation. Раскидал по папкам.
