+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="ru.androidschool.intensiv.MainActivity">
+        <activity android:name="ru.androidschool.intensiv.presentation.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
