--- app/src/main/java/ru/androidschool/intensiv/data/MockRepository.kt
@@ -1,3 +1,0 @@
-package ru.androidschool.intensiv.data
-
-object MockRepository
