+++ app/src/main/java/ru/androidschool/intensiv/data/network/MovieApiClient.kt
@@ -1,0 +1,31 @@
+package ru.androidschool.intensiv.data.network
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
+import ru.androidschool.intensiv.BuildConfig
+
+object MovieApiClient {
+
+    private var client: OkHttpClient =
+        if (BuildConfig.DEBUG) {
+            OkHttpClient.Builder()
+                .addInterceptor(HttpLoggingInterceptor().apply { this.level = HttpLoggingInterceptor.Level.BODY })
+                .build()
+        } else {
+            OkHttpClient.Builder().build()
+        }
+
+    val apiClient: MovieApiInterface by lazy {
+        val retrofit = Retrofit.Builder()
+            .baseUrl(BuildConfig.BASE_URL)
+            .client(client)
+            .addCallAdapterFactory(RxJava2CallAdapterFactory.create()) // обязательно адаптер для RxJava
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+
+        return@lazy retrofit.create(MovieApiInterface::class.java)
+    }
+}
