+++ app/src/main/java/ru/androidschool/intensiv/data/repository/MockRepository.kt
@@ -1,0 +1,3 @@
+package ru.androidschool.intensiv.data.repository
+
+object MockRepository
