--- app/src/main/java/ru/androidschool/intensiv/database/MovieFinderAppTypeConverters.kt
@@ -1,16 +1,0 @@
-package ru.androidschool.intensiv.database
-
-import androidx.room.TypeConverter
-import ru.androidschool.intensiv.utils.ViewFeature
-
-class MovieFinderAppTypeConverters {
-    @TypeConverter
-    fun toViewFeature(stringViewFeature: String): ViewFeature {
-        return ViewFeature.valueOf(stringViewFeature)
-    }
-
-    @TypeConverter
-    fun fromViewFeature(viewFeature: ViewFeature): String {
-        return viewFeature.name
-    }
-}
