+++ app/src/main/java/ru/androidschool/intensiv/presentation/feed/MovieItem.kt
@@ -1,0 +1,27 @@
+package ru.androidschool.intensiv.presentation.feed
+
+import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
+import com.xwray.groupie.kotlinandroidextensions.Item
+import kotlinx.android.synthetic.main.item_with_text.*
+import ru.androidschool.intensiv.R
+import ru.androidschool.intensiv.data.vo.MovieVO
+import ru.androidschool.intensiv.ui.loadImage
+
+class MovieItem(
+    private val content: MovieVO,
+    private val onClick: (movie: MovieVO) -> Unit
+) : Item() {
+
+    override fun getLayout() = R.layout.item_with_text
+
+    override fun bind(viewHolder: GroupieViewHolder, position: Int) {
+        viewHolder.description.text = content.title
+        viewHolder.movie_rating.rating = content.rating
+        viewHolder.content.setOnClickListener {
+            onClick.invoke(content)
+        }
+
+        // TODO Получать из модели
+        viewHolder.image_preview.loadImage(content.getPosterPathWithImageUrl())
+    }
+}
