+++ app/src/main/res/layout/feed_header.xml
@@ -52,7 +52,7 @@
                         android:layout_height="wrap_content"
                         android:orientation="vertical">
 
-                        <ru.androidschool.intensiv.ui.SearchBar
+                        <ru.androidschool.intensiv.presentation.search.SearchBar
                             android:id="@+id/search_toolbar"
                             android:layout_width="match_parent"
                             android:layout_height="@dimen/search_bar_height"
