+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,13 +4,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".ui.profile.ProfileFragment">
+    tools:context=".presentation.profile.ProfileFragment">
 
     <LinearLayout xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".MainActivity">
+        tools:context=".presentation.MainActivity">
 
         <RelativeLayout
             android:layout_width="match_parent"
