+++ app/src/main/res/layout/fragment_watchlist.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.watchlist.WatchlistFragment">
+    tools:context=".presentation.watchlist.WatchlistFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/movies_recycler_view"
