+++ app/src/main/res/layout/tv_shows_fragment.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/mainFeedBackground"
-    tools:context="ru.androidschool.intensiv.ui.tvshows.TvShowsFragment">
+    tools:context=".presentation.tvshows.TvShowsFragment">
 
     <include layout="@layout/progress_bar" />
