+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -5,7 +5,7 @@
     app:startDestination="@+id/home_dest">
     <fragment
         android:id="@+id/home_dest"
-        android:name="ru.androidschool.intensiv.ui.feed.FeedFragment"
+        android:name="ru.androidschool.intensiv.presentation.feed.FeedFragment"
         android:label="@string/home"
         tools:layout="@layout/feed_fragment">
 
@@ -13,7 +13,7 @@
 
     <fragment
         android:id="@+id/flow_step_one_dest"
-        android:name="ru.androidschool.intensiv.ui.movie_details.MovieDetailsFragment"
+        android:name="ru.androidschool.intensiv.presentation.movie_details.MovieDetailsFragment"
         tools:layout="@layout/movie_details_fragment_part">
         <argument
             android:name="flowStepNumber"
@@ -27,7 +27,7 @@
 
     <fragment
         android:id="@+id/flow_step_two_dest"
-        android:name="ru.androidschool.intensiv.ui.movie_details.MovieDetailsFragment"
+        android:name="ru.androidschool.intensiv.presentation.movie_details.MovieDetailsFragment"
         tools:layout="@layout/movie_details_fragment_part">
 
         <argument
@@ -42,7 +42,7 @@
 
     <fragment
         android:id="@+id/tv_shows_dest"
-        android:name="ru.androidschool.intensiv.ui.tvshows.TvShowsFragment"
+        android:name="ru.androidschool.intensiv.presentation.tvshows.TvShowsFragment"
         android:label="@string/tv_show"
         tools:layout="@layout/tv_shows_fragment">
 
@@ -52,21 +52,21 @@
     </fragment>
     <fragment
         android:id="@+id/movie_details_fragment"
-        android:name="ru.androidschool.intensiv.ui.movie_details.MovieDetailsFragment"
+        android:name="ru.androidschool.intensiv.presentation.movie_details.MovieDetailsFragment"
         android:label="details_fragment"
         tools:layout="@layout/movie_details_fragment_part" />
     <fragment
         android:id="@+id/profile_fragment"
-        android:name="ru.androidschool.intensiv.ui.profile.ProfileFragment"
+        android:name="ru.androidschool.intensiv.presentation.profile.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
     <fragment
         android:id="@+id/search_dest"
-        android:name="ru.androidschool.intensiv.ui.search.SearchFragment"
+        android:name="ru.androidschool.intensiv.presentation.search.SearchFragment"
         android:label="SearchFragment" />
 
     <fragment
         android:id="@+id/tv_show_details_fragment"
-        android:name="ru.androidschool.intensiv.ui.tvshows.TvShowDetailsFragment"
+        android:name="ru.androidschool.intensiv.presentation.tvshows.TvShowDetailsFragment"
         android:label="TvShowDetailsFragment" />
 </navigation>
